initial commit
contact form; the version ending in "-html5" is the baseline; the js is
actually commented out....either switch the alert()s to messages or
write a different script. on submission success, displays a thank you
message to the user and submits an email to me populated with the form's
details.